+++ core/src/main/scala/com/krux/hyperion/cli/AwsAction.scala
@@ -1,12 +1,13 @@
 package com.krux.hyperion.cli
 
-import com.krux.hyperion.{ DataPipelineDef, HyperionAwsClient }
+import com.krux.hyperion.DataPipelineDefGroup
+import com.krux.hyperion.client.{ AwsClientForDef, AwsClient }
 
 private[hyperion] trait AwsAction extends Action {
 
-  def apply(options: Options, client: HyperionAwsClient): Boolean
+  def apply(options: Options, client: AwsClientForDef): Boolean
 
-  def apply(options: Options, pipelineDef: DataPipelineDef): Boolean =
-    apply(options, HyperionAwsClient(pipelineDef, options.region, options.roleArn))
+  def apply(options: Options, pipelineDef: DataPipelineDefGroup): Boolean =
+    apply(options, AwsClient(pipelineDef, options.region, options.roleArn))
 
 }