+++ examples/src/test/scala/com/krux/hyperion/examples/ExampleRedshiftLoadSpec.scala
@@ -4,6 +4,7 @@ import org.scalatest.WordSpec
 import org.json4s.jackson.JsonMethods._
 import org.json4s.JsonDSL._
 import org.json4s._
+
 import com.krux.hyperion.DataPipelineDef._